Initial draft of tabbed preferences dialog
