+++ netlogo-gui/resources/i18n/GUI_Strings_en.properties
@@ -254,10 +254,6 @@ tabs.code.includedFiles.new = New Source File
 tabs.code.includedFiles.open = Open Source File
 tabs.code.separateCodeWindow = Separate code window
 tabs.code.preferences = Preferences
-tabs.code.preferences.title = Code Tab Preferences
-tabs.code.preferences.message = <html>The following global preferences apply to the main Code Tab<br>as well as any included file tabs and LevelSpace code tabs.</html>
-tabs.code.indentAutomatically = Auto indent
-tabs.code.editorLineNumbers = Show line numbers
 
 tabs.external.new = New File {0}
 tabs.external.close = Close
@@ -566,6 +562,7 @@ resource.noResource = Resource "{0}" does not exist.
 # tools
 
 tools.preferences = NetLogo Preferences
+tools.preferences.general = General
 tools.preferences.restartRequired = (restart required)
 tools.preferences.reloadRequired = (model reload required)
 tools.preferences.uiLanguage = User Interface Language:
@@ -574,12 +571,6 @@ tools.preferences.logDirectory = Directory to store logs:
 tools.preferences.logEvents = Events to log:
 tools.preferences.loadLastOnStartup = Reload the last used model on startup:
 tools.preferences.reloadOnExternalChanges = Reload model on external changes:
-tools.preferences.includedFilesMenu = Always show "Included Files" menu (in Code tab):
-tools.preferences.proceduresMenuSortOrder = Sort order of procedures menu (in Code tab):
-tools.preferences.proceduresSortAlphabetical = Alphabetical
-tools.preferences.proceduresSortByOrderOfAppearance = Order of appearance
-tools.preferences.focusOnError = Auto-focus text or tab with errors when model is compiled
-tools.preferences.startSeparateCodeTab = Open code tab in separate window on startup
 tools.preferences.emptyDirectory = Logging was enabled, but no log directory was specified.
 tools.preferences.badPermissions = NetLogo does not have permission to write to the specified log directory.
 tools.preferences.missingDirectory = The specified log directory does not exist, would you like to create it?
@@ -587,11 +578,22 @@ tools.preferences.boldWidgetNames = Bold widget names
 tools.preferences.uiScale = UI Scale
 tools.preferences.scaleError = UI Scale must be a valid floating point number.
 
-tools.themes = Color Themes
-tools.themes.text = These themes control the colors of the entire application,<br>including toolbars, menus, dialogs, and widgets.
-tools.themes.classic = Classic
-tools.themes.light = Light
-tools.themes.dark = Dark
+tools.preferences.code = Code
+tools.preferences.code.message = <html>The following global preferences apply to the main Code Tab<br>as well as any included file tabs and LevelSpace code tabs.</html>
+tools.preferences.includedFilesMenu = Always show "Included Files" menu (in Code tab):
+tools.preferences.proceduresMenuSortOrder = Sort order of procedures menu (in Code tab):
+tools.preferences.proceduresSortAlphabetical = Alphabetical
+tools.preferences.proceduresSortByOrderOfAppearance = Order of appearance
+tools.preferences.focusOnError = Auto-focus text or tab with errors when model is compiled
+tools.preferences.startSeparateCodeTab = Open code tab in separate window on startup
+tools.preferences.indentAutomatically = Auto indent
+tools.preferences.editorLineNumbers = Show line numbers
+
+tools.preferences.themes = Color Themes
+tools.preferences.themes.text = These themes control the colors of the entire application,<br>including toolbars, menus, dialogs, and widgets.
+tools.preferences.themes.classic = Classic
+tools.preferences.themes.light = Light
+tools.preferences.themes.dark = Dark
 
 tools.libraries = Extensions
 tools.libraries.categories.extensions = Extensions
